requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
astroid=1.6.5=py36_0
blas=1.0=mkl
bzip2=1.0.6=h9a117a8_4
ca-certificates=2019.1.23=0
cairo=1.14.12=h8948797_3
certifi=2018.11.29=py36_0
cffi=1.11.5=py36he75722e_1
cuda100=1.0=0
cuda80=1.0=h205658b_0
cudatoolkit=8.0=3
cycler=0.10.0=py36h93f1223_0
dbus=1.13.2=h714fa37_1
decorator=4.2.1=py36_0
expat=2.2.5=he0dffb1_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.9.1=h8a8886c_1
glib=2.56.2=hd408876_0
graphite2=1.3.13=h23475e2_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=h9c2bf20_1
intel-openmp=2018.0.0=hc7b2577_8
ipykernel=4.8.0=py36_0
ipython=6.2.1=py36h88c514a_1
ipython_genutils=0.2.0=py36hb52b0d5_0
isort=4.3.4=py36_0
jasper=2.0.14=h07fcdf6_1
jedi=0.11.1=py36_0
jpeg=9b=h024ee3a_2
jupyter_client=5.2.2=py36_0
jupyter_core=4.4.0=py36h7c827e3_0
kiwisolver=1.0.1=py36hf484d3e_0
lazy-object-proxy=1.3.1=py36h10fcdad_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.2.0=h9f7466a_2
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3=h7b6447c_0
libpng=1.6.36=hbc83047_0
libprotobuf=3.6.1=hd408876_0
libsodium=1.0.15=hf101ebd_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.10=h2733197_2
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.9=he19cac6_0
matplotlib=2.2.2=py36hb69df0a_2
mccabe=0.6.1=py36h5ad9710_1
mkl=2018.0.1=h19d6760_4
ncurses=6.1=he6710b0_1
ninja=1.8.2=py36h6bb024c_1
numpy=1.14.2=py36hdbf6ddf_0
olefile=0.46=py36_0
opencv=3.4.2=py36h6fd60c2_1
openssl=1.1.1a=h7b6447c_0
parso=0.1.1=py36h35f843b_0
pcre=8.42=h439df22_0
pexpect=4.3.1=py36_0
pickleshare=0.7.4=py36h63277f8_0
pillow=5.4.1=py36h34e0f95_0
pip=9.0.1=py36h6c6f9ce_4
pixman=0.36.0=h7b6447c_0
prompt_toolkit=1.0.15=py36h17d85b1_0
ptyprocess=0.5.2=py36h69acd42_0
py-opencv=3.4.2=py36hb342d67_1
pycparser=2.19=py36_0
pygments=2.2.0=py36h0d3125c_0
pylint=1.9.2=py36_0
pyparsing=2.2.0=py36hee85983_1
pyqt=5.6.0=py36h0386399_5
python=3.6.8=h0371630_0
python-dateutil=2.6.1=py36h88d3b88_1
pytorch=1.0.0=py3.6_cuda10.0.130_cudnn7.4.1_1
pytz=2018.3=py36_0
pyzmq=16.0.3=py36he2533c7_0
qt=5.6.3=h8bf5577_3
readline=7.0=ha6073c6_4
setuptools=38.4.0=py36_0
simplegeneric=0.8.1=py36_2
sip=4.18.1=py36h51ed4ed_2
six=1.11.0=py36h372c433_1
sqlite=3.26.0=h7b6447c_0
tk=8.6.8=hbc83047_0
torchvision=0.2.1=py_2
tornado=4.5.3=py36_0
traitlets=4.3.2=py36h674d592_0
wcwidth=0.1.7=py36hdf4376a_0
wheel=0.30.0=py36hfd4bba0_1
wrapt=1.10.11=py36h28b7045_0
xz=5.2.4=h14c3975_4
zeromq=4.2.2=hbedb6e5_2
zlib=1.2.11=ha838bed_2
zstd=1.3.7=h0b5b093_0
